pyproject.toml
[project]
name = "llm-security"
version = "0.1.0"
description = "Demonstrates 9 llm security layers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "llm-guard>=0.3.16",
    "redis>=5.2.0",
    "chromadb>=0.5.0",
    "openai>=1.60.0",
    "pyjwt>=2.10.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.9.0",
    "httpx>=0.28.0",
    "passlib>=1.7.4",
    "bcrypt>=4.2.0",
]
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",

]
